+++ src/quacc/recipes/dftb/_base.py
@@ -12,6 +12,7 @@ from quacc.utils.dicts import merge_dicts
 from quacc.utils.files import check_logfile
 
 if TYPE_CHECKING:
+    from pathlib import Path
     from typing import Any
 
     from ase.atoms import Atoms
@@ -27,7 +28,7 @@ def base_fn(
     calc_defaults: dict[str, Any] | None = None,
     calc_swaps: dict[str, Any] | None = None,
     additional_fields: dict[str, Any] | None = None,
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
 ) -> RunSchema:
     """
     Base job function for DFTB+ recipes.
@@ -46,7 +47,7 @@ def base_fn(
     additional_fields
         Any additional fields to supply to the summarizer.
     copy_files
-        Files to copy to the runtime directory.
+        File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
 
     Returns
     -------