+++ src/quacc/recipes/vasp/mp.py
@@ -36,7 +36,7 @@ def mp_prerelax_job(
     atoms: Atoms,
     preset: str | None = "MPScanSet",
     bandgap: float | None = None,
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
     **calc_kwargs,
 ) -> VaspSchema:
     """
@@ -52,7 +52,7 @@ def mp_prerelax_job(
     bandgap
         Estimate for the bandgap in eV.
     copy_files
-        Files to copy to the runtime directory.
+        File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **calc_kwargs
         Custom kwargs for the Vasp calculator. Set a value to
         `None` to remove a pre-existing key entirely. For a list of available
@@ -86,7 +86,7 @@ def mp_relax_job(
     atoms: Atoms,
     preset: str | None = "MPScanSet",
     bandgap: float | None = None,
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
     **calc_kwargs,
 ) -> VaspSchema:
     """
@@ -102,7 +102,7 @@ def mp_relax_job(
     bandgap
         Estimate for the bandgap in eV.
     copy_files
-        Files to copy to the runtime directory.
+        File(s) to copy to the runtime directory. If a directory is provided, it will be recursively unpacked.
     **calc_kwargs
         Dictionary of custom kwargs for the Vasp calculator. Set a value to
         `None` to remove a pre-existing key entirely. For a list of available