+++ src/quacc/runners/ase.py
@@ -21,6 +21,7 @@ except ImportError:
     Sella = None
 
 if TYPE_CHECKING:
+    from pathlib import Path
     from typing import Any, TypedDict
 
     from ase.atoms import Atoms
@@ -37,7 +38,9 @@ if TYPE_CHECKING:
 
 
 def run_calc(
-    atoms: Atoms, geom_file: str | None = None, copy_files: list[str] | None = None
+    atoms: Atoms,
+    geom_file: str | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
 ) -> Atoms:
     """
     Run a calculation in a scratch directory and copy the results back to the original
@@ -112,7 +115,7 @@ def run_opt(
     optimizer: Optimizer = FIRE,
     optimizer_kwargs: OptimizerKwargs | None = None,
     run_kwargs: dict[str, Any] | None = None,
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
 ) -> Optimizer:
     """
     Run an ASE-based optimization in a scratch directory and copy the results back to
@@ -192,7 +195,7 @@ def run_opt(
 def run_vib(
     atoms: Atoms,
     vib_kwargs: VibKwargs | None = None,
-    copy_files: list[str] | None = None,
+    copy_files: str | Path | list[str | Path] | None = None,
 ) -> Vibrations:
     """
     Run an ASE-based vibration analysis in a scratch directory and copy the results back