+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.4]
+
+### Added
+
+- Added core Quantum ESPRESSO recipes and custom calculator. (@tomdemeyere)
+- Added a new function, `quacc.utils.files.copy_decompress_files_from_dir`
+
+### Changed
+
+- The `copy_files` kwarg in most recipes can now include a directory
+- Renamed `quacc.utils.files.copy_decompress` to `quacc.utils.files.copy_decompress_files`
+
+### Fixed
+
+- Quacc settings that are paths now have `~/` expanded in most cases
+
 ## [0.4.3]
 
 ### Added